Fixed P1 Öffentliche Registrierung

+++ server/src/controllers/user.controller.ts
@@ -52,6 +52,10 @@ export const updateUserById = async (request: Request, response: Response) => {
 /* ========== DELETE ========== */
 
 export const deleteUser = async (request: Request, response: Response) => {
+    if (request.user?.id === request.params.id) {
+        throw new AppException("You cannot delete your own account!", 403, "USER_DELETE_SELF_FORBIDDEN");
+    }
+
     await userService.deleteUser(request.params.id as string);
 
     return response.status(200).json({
